feat(dev): add .editorconfig, set vscode indentation
The main motivation for this change is that PyCharm does not correctly detect
two-space indentation in tsx files, which is very frustrating (backspace
deletes four spaces, then you need to type two more, etc). Along with the
vscode change, it is also useful for setting the right indentation in new
files.

Indentation size and type can often be controversial. This editorconfig only
codifies the existing convention of 2 spaces for frontend and 4 spaces for
backend. No indentation size is set for html, bash, lua, makefile, etc, which
currently use a mix of 2 and 4 spaces (except makefile).

Some existing files do not match these rules. They mostly fall into three
categories:

1. Copied and pasted string blobs (mostly JSON pasted into tests).
2. Files which are not normally edited with text editors, e.g. raster images.
3. Mistakes.

Case 1 should either be reformatted as the files are edited in future commits,
or left the way they are.

Case 2 shouldn't pose any issues.

Case 3 should probably be dealt with on a case-by-case basis in future commits.